Revert "The profile tab strip stops being rebuilt by the navigation it performs (#649)"
This reverts commit da4e2d8d546d5712026a32b2c0a041f61e9a5158.
